+++ lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/query_parameters.dart
@@ -1,5 +1,5 @@
 export 'close_approach_body.dart';
 export 'data_output.dart';
 export 'sbdb_cad_query_parameters.dart';
-export 'small_body.dart';
+export 'small_body_filter.dart';
 export 'small_body_selector.dart';