+++ lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/sbdb_cad_query_parameters.dart
@@ -6,7 +6,7 @@ import 'package:hrk_batteries/hrk_batteries.dart';
 import '../../../../../globals.dart';
 import 'close_approach_body.dart';
 import 'data_output.dart';
-import 'small_body.dart';
+import 'small_body_filter.dart';
 import 'small_body_selector.dart';
 
 part 'sbdb_cad_query_parameters.freezed.dart';
@@ -61,7 +61,7 @@ class SbdbCadQueryParameters with _$SbdbCadQueryParameters {
     value: 0.05,
     unit: distUnitDefault,
   );
-  static const SmallBody smallBodyDefault = SmallBody.neo;
+  static const SmallBodyFilter smallBodyFilterDefault = SmallBodyFilter.neo;
   static const CloseApproachBody closeApproachBodyDefault =
       CloseApproachBody.earth;
 
@@ -114,24 +114,24 @@ class SbdbCadQueryParameters with _$SbdbCadQueryParameters {
     return queryparameters;
   }
 
-  SbdbCadQueryParameters copyWithSmallBody(
-    SmallBody smallBody, {
+  SbdbCadQueryParameters copyWithSmallBodyFilter(
+    SmallBodyFilter smallBodyFilter, {
     bool? value = true,
   }) {
     // TODO(hrishikesh-kadam): unused_result lint is false positive
     // Issue: https://github.com/dart-lang/sdk/issues/52314
     // Remove the lint after the following fix is in stable
     // https://github.com/dart-lang/sdk/commit/b0b393773fa7d7965c6c9ea3a501e93dc60812f9
-    return switch (smallBody) {
-      smallBodyDefault => this,
+    return switch (smallBodyFilter) {
+      smallBodyFilterDefault => this,
       // ignore: unused_result
-      SmallBody.pha => copyWith(pha: value),
+      SmallBodyFilter.pha => copyWith(pha: value),
       // ignore: unused_result
-      SmallBody.nea => copyWith(nea: value),
+      SmallBodyFilter.nea => copyWith(nea: value),
       // ignore: unused_result
-      SmallBody.comet => copyWith(comet: value),
+      SmallBodyFilter.comet => copyWith(comet: value),
       // ignore: unused_result
-      SmallBody.neaComet => copyWith(neaComet: value),
+      SmallBodyFilter.neaComet => copyWith(neaComet: value),
     };
   }