+++ test/services/ssd_cneos/sbdb_cad_api/sbdb_cad_api_test.dart
@@ -79,8 +79,9 @@ void main() {
         });
       });
 
-      test('$SmallBody', () async {
-        queryParameters = queryParameters.copyWithSmallBody(SmallBody.pha);
+      test('$SmallBodyFilter', () async {
+        queryParameters =
+            queryParameters.copyWithSmallBodyFilter(SmallBodyFilter.pha);
         final Response<SbdbCadBody> response = await api.get(
           queryParameters: queryParameters.toJson(),
         );